Bug fix: syntastic: address issue finding .jscsrc
The previous declaration would attempt to initialize the javascript
checkers at vim start time. $CWD would be assign the value
of wherever vim was launched and led to not finding a .jscsrc file
in local project directories.

At the cost of re-declaring the syntastic checker when a buffer
is created when opening javascript files for the first time, the
syntastic checkers are now accurately assigned.